Add support for multiple modules
