+++ src/main/kotlin/com/emberjs/utils/AnActionEventExtensions.kt
@@ -1,22 +1,19 @@
 package com.emberjs.utils
 
-import com.emberjs.project.EmberProjectComponent
 import com.intellij.openapi.actionSystem.AnActionEvent
 import com.intellij.openapi.actionSystem.LangDataKeys
+import com.intellij.openapi.module.Module
+import com.intellij.openapi.roots.ModuleRootManager
 import com.intellij.openapi.vfs.VfsUtil.isAncestor
 import com.intellij.openapi.vfs.VirtualFile
 
-
-val AnActionEvent.virtualFile: VirtualFile?
-    get() = getData(LangDataKeys.VIRTUAL_FILE)
+val AnActionEvent.module: Module?
+    get() = getData(LangDataKeys.MODULE)
 
 val AnActionEvent.emberRoot: VirtualFile?
     get() {
-        val project = project ?: return null
-        val file = virtualFile ?: return null
-        val roots = EmberProjectComponent.getInstance(project)?.roots ?: return null
-
-        return roots.filter { isAncestor(it, file, true) }.firstOrNull()
+        val module = module ?: return null
+        return ModuleRootManager.getInstance(module).contentRoots.find { it.isEmberFolder }
     }
 
 val AnActionEvent.hasEmberRoot: Boolean
